did the requested changes
