+++ src/digitalIdentity-frontend/src/app/pages/login-page/login-page.component.ts
@@ -2,7 +2,7 @@ import { Component, isDevMode, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { HttpParams } from '@angular/common/http';
 import { MatDialog } from '@angular/material/dialog';
-import { InformationPopUpComponent } from '../../shared/information-pop-up/information-pop-up.component';
+import { InformationPopUpComponent } from '../../shared/pop-up/information-pop-up/information-pop-up.component';
 import { Router } from '@angular/router';
 import { BackendHttpService } from 'src/app/services/backend-http-service/backend-http-service.service';
 @Component({
