+++ src/digitalIdentity-frontend/src/app/pages/schema/schema-overview-page/schema-overview-page.component.html
