+++ src/digitalIdentity-frontend/src/app/pages/schema/schema-overview-page/schema-overview-page.component.ts
@@ -3,7 +3,7 @@ import { FormControl, FormGroup } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { MatTableDataSource } from '@angular/material/table';
 import { BackendHttpService } from 'src/app/services/backend-http-service/backend-http-service.service';
-import { InformationPopUpComponent } from '../../../shared/information-pop-up/information-pop-up.component';
+import { InformationPopUpComponent } from 'src/app/shared/pop-up/information-pop-up/information-pop-up.component';
 export interface attributeType {
   name: string;
   type: 'string' | 'date' | 'number';
@@ -17,9 +17,9 @@ export interface schemaDataType {
 }
 
 @Component({
-  selector: 'app-schema-overview',
-  templateUrl: './schema-overview.component.html',
-  styleUrls: ['./schema-overview.component.css'],
+  selector: 'app-schema-overview-page',
+  templateUrl: './schema-overview-page.component.html',
+  styleUrls: ['./schema-overview-page.component.css'],
 })
 export class SchemaOverviewComponent implements OnInit {
   constructor(
