--- src/digitalIdentity-frontend/src/app/pages/settings/settings.component.html
@@ -1,1 +1,0 @@
-<p>admin settings work!</p>
