--- src/digitalIdentity-frontend/src/app/pages/settings/settings.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-setting',
-  templateUrl: './settings.component.html',
-  styleUrls: ['./settings.component.css'],
-})
-export class SettingsComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
