+++ src/digitalIdentity-frontend/src/app/shared/pop-up/information-pop-up/information-pop-up.component.spec.ts
